Create perf_breaker.yml
Moving this bad boy back to master. The purpose of this work flow is to check that all algorithms in this repo were not broken by a recent change.

This workflow only checks that a certain reward has been reached by an algorithm on a specific env. I plan to add config files for as many combinations of algos/envs as possible. This will be a requirement in the future. This way, we can always check how performance is doing with big updates.